+++ dart_json_schema_annotations/lib/dart_json_schema_annotations.dart
@@ -1,2 +1,1 @@
 export 'package:dart_json_schema_annotations/src/field.dart';
-export 'package:dart_json_schema_annotations/src/typed_fields.dart';